copy commands now respect mtime

+++ code/tools/convert.py
@@ -5,14 +5,35 @@ from typing import List
 
 from code import util
 
+LODs = ["-1000", "-500", "-250"]
+
+
+def copy_single(source_path: pathlib.Path, target_path: pathlib.Path) -> None:
+    if target_path.exists() and target_path.stat().st_mtime > source_path.stat().st_mtime:
+        return
+
+    target_path.parent.mkdir(parents=True, exist_ok=True)
+    shutil.copy(source_path, target_path)
+
 
 def copy(source: str, target: str, path: str, target_fn: str = None) -> None:
     if target_fn is None:
         target_fn = path
     source_path = pathlib.Path(source, path)
     target_path = pathlib.Path(target, target_fn)
-    target_path.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copy(source_path, target_path)
+    copy_single(source_path, target_path)
+
+    for lod in LODs:
+        if lod not in path:
+            continue
+
+        for other_lod in LODs:
+            if other_lod == lod:
+                continue
+
+            new_source = str(source_path).replace(lod, other_lod)
+            new_target = str(target_path).replace(lod, other_lod)
+            copy_single(pathlib.Path(new_source), pathlib.Path(new_target))
 
 
 def patch_egg(ctx: util.Context, path: str) -> None:
@@ -46,19 +67,25 @@ def eggtrans(ctx: util.Context, path: str) -> None:
 def copy_errors(ctx: util.Context, path: str, errored_files: List[str]) -> bool:
     partial_abspath = ctx.resources_path + "/" + "/".join(path.split("/")[:-1])
     for x in errored_files:
+        target_path = f"{partial_abspath}/{x}"
+        possible_paths = []
         if os.path.exists(f"{ctx.resources_path}/{x}"):
-            p2 = f"{ctx.resources_path}/{x}"
-        elif os.path.exists(f"{partial_abspath}/{x}"):
-            p2 = f"{partial_abspath}/{x}"
-        else:
+            possible_paths.append(f"{ctx.resources_path}/{x}")
+        if os.path.exists(f"{partial_abspath}/{x}"):
+            possible_paths.append(f"{partial_abspath}/{x}")
+
+        if not possible_paths:
             p2 = input(f"Unable to find {x} in resources. Enter path to copy from: ")
             if not p2:
                 print(f"Unable to get {x} from anywhere, aborting.")
                 return False
-        print(f"Copying {x} from resources.")
-        target_path = p2.replace(ctx.resources_path, ctx.working_path)
-        pathlib.Path(target_path).parent.mkdir(exist_ok=True, parents=True)
-        shutil.copy(p2, target_path)
+            possible_paths.append(p2)
+
+        newest_file = max(possible_paths, key=os.path.getctime)
+        if newest_file != target_path:
+            print(f"Copying {x} from resources.")
+            pathlib.Path(target_path).parent.mkdir(exist_ok=True, parents=True)
+            shutil.copy(newest_file, target_path)
 
     return True
 
@@ -103,7 +130,7 @@ def build_lods(ctx: util.Context, path: str) -> None:
     if not path.endswith(".egg") and not path.endswith(".bam"):
         raise Exception("Only .egg and .bam files can be triplicated!")
 
-    if any(f"-{lod}." in path for lod in (1000, 500, 250)):
+    if any(f"{lod}." in path for lod in LODs):
         raise Exception("This file is already triplicated!")
 
     abspath = pathlib.Path(path)
@@ -112,6 +139,6 @@ def build_lods(ctx: util.Context, path: str) -> None:
 
     extension = path[-3:]
     base_name = path[:-4]
-    target_names = [f"{base_name}-{lod}.{extension}" for lod in (1000, 500, 250)]
+    target_names = [f"{base_name}{lod}.{extension}" for lod in LODs]
     for target_name in target_names:
         copy(ctx.working_path, ctx.working_path, path, target_name)
